// src/api/models/banner.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */
import type { BannerImage } from './bannerImage';

export interface Banner {
  id: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  title_en?: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  description_en?: string | null;
  /** @nullable */
  image: BannerImage;
  display_order: number;
  is_active: boolean;
  created_at: string;
  /** @nullable */
  created_by?: string | null;
  updated_at: string;
  /** @nullable */
  updated_by?: string | null;
}